% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mesh.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot mesh}
\usage{
plot_mesh(mesh_obj, data_obj, catch_var = "catch_weight", group = "Total")
}
\arguments{
\item{mesh_obj}{Mesh object from sdmTMB}

\item{data_obj}{Raw data}

\item{catch_var}{Variable name in " " to be plotted}

\item{group}{Title for plot}
}
\description{
Plot mesh
}
