% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine-model-functions.R
\name{refine_cond_model}
\alias{refine_cond_model}
\title{refine non-delta model; particularly one with anisotropy and/or without shared ranges}
\usage{
refine_cond_model(m, set_formula = cond_formula, dist = knot_distance)
}
\description{
refine non-delta model; particularly one with anisotropy and/or without shared ranges
}
