% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-predicted-distribution.R
\name{trim_predictions}
\alias{trim_predictions}
\title{Trim predictions to represent a percentage of the mean total abundance}
\usage{
trim_predictions(data, prop_threshold = 0.01)
}
\arguments{
\item{data}{A grid in which cells must be defined by "X", "Y", time variable is "year",
and biomass/abundance stored as "density"}

\item{prop_threshold}{Default is 0.01, which keeps 99% of the predicted abundance}
}
\description{
Trim predictions to represent a percentage of the mean total abundance
}
