% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-predicted-distribution.R
\name{trim_predictions_by_year}
\alias{trim_predictions_by_year}
\title{Trim predictions separately for each year}
\usage{
trim_predictions_by_year(data, prop_threshold = 0.01)
}
\description{
Trim predictions separately for each year
}
